new_branch commit 4

+++ app/src/main/java/com/hfad/allweather/data/remote/dto/forecast_weatherDto/CurrentForecastWeatherDto.kt
@@ -13,7 +13,7 @@ data class CurrentForecastWeatherDto(
     val is_day: Int,
     val last_updated: String,
     val last_updated_epoch: Int,
-    val precip_in: Int,
+    val precip_in: Double,
     val precip_mm: Double,
     val pressure_in: Double,
     val pressure_mb: Int,
